TsCanCode 扫描问题修改 Signed-off-by: wangyunkai <wangyunkai1@h-******>
# message auto-generated for no-merge-commit merge:
merge tscancode into master

修复TscanCode工具扫描出的问题

Created-by: insightDeveloper

Match-id-0d2539c86bb2bd7a48f0d23fa74faa26668501bb

+++ server/src/modules/base/core/JsonFileProcess.cpp
@@ -102,12 +102,12 @@ bool JsonFileProcess::SeekCharPosition(std::ifstream &file, char c)
     }
     file.seekg(cur);
     std::string str(buffer.get(), readCount);
-    uint64_t offset = str.find(c);
+    size_t offset = str.find(c);
     if (offset == std::string::npos) {
         Dic::Server::ServerLog::Error("Failed to find separator.");
         return false;
     }
-    file.seekg(offset, std::ifstream::cur);
+    file.seekg(static_cast<int64_t>(offset), std::ifstream::cur);
     return true;
 }
